Update README.md
For our programming environment we used the Jupyter Notebook with Python language interpreter, as well as Jetbrains Datalore in order to collaborate. Our main computational tool used to approximate the integrals for each periodic Fourier coefficient was the Simpson's rule.   

To implement the Simpson's rule in python we used the following algorithm

Initialize specification variables
Define a function to integrate
Split the integration interval into n steps of equal length
Sum over the even and odd indices of the integral n times 
Return the weighted average as an approximation for the integral